Add endpoint for GET /revisions.
