+++ deckhand/control/api.py
@@ -21,6 +21,7 @@ from oslo_log import log as logging
 from deckhand.conf import config
 from deckhand.control import base as api_base
 from deckhand.control import documents
+from deckhand.control import revision_documents
 from deckhand.control import revisions
 from deckhand.control import secrets
 from deckhand.db.sqlalchemy import api as db_api
@@ -69,7 +70,9 @@ def start_api(state_manager=None):
 
     v1_0_routes = [
         ('documents', documents.DocumentsResource()),
-        ('revisions/{revision_id}/documents', revisions.RevisionsResource()),
+        ('revisions', revisions.RevisionsResource()),
+        ('revisions/{revision_id}/documents',
+         revision_documents.RevisionDocumentsResource()),
         ('secrets', secrets.SecretsResource())
     ]
